requirements.txt
-- prefer-binaries
Flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
waitress==2.1.2
concurrent-log-handler==0.24.0
numpy==1.24.3
